Add missed test for getting, listing and generating upload Ids
