--- server/src/test/java/umm3601/plant/TestListUploadIds.java
@@ -1,25 +1,0 @@
-package umm3601.plant;
-
-import org.junit.Before;
-import org.junit.Test;
-import umm3601.digitalDisplayGarden.PlantController;
-
-import java.io.IOException;
-
-public class TestListUploadIds {
-
-    private final static String databaseName = "data-for-testing-only";
-    private PlantController plantController;
-
-    @Before
-    public void populateDB() throws IOException {
-        PopulateMockDatabase db = new PopulateMockDatabase();
-        db.clearAndPopulateDBAgain();
-        plantController = new PlantController(databaseName);
-    }
-
-    @Test
-    public void listingOfUploadIds() throws IOException {
-
-    }
-}
